+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.24.0-wip
+
+- **BREAKING CHANGE**: `SANDBOX_OUTPUT` environment variable passed to `--sandbox-runner` is now a JSON-encoded list of directory paths instead of a colon-separated string.
+- Fix Windows failure by only populating sandbox environment variables when `--sandbox-runner` is specified (#1633).
+
 ## 0.23.19
 
 - Include hosted and SDK `dev_dependencies` during static analysis and documentation generation.